# docs/v2.swagger.yaml
openapi: 3.0.3
info:
  title: OrdFS Server API - V2
  description: |
    API server for serving Bitcoin ordinal inscriptions and OrdFS content.

    Supports:
    - Content retrieval by txid or outpoint
    - Version tracking for ordinal reinscriptions
    - OrdFS directory navigation
    - DNS-based custom domain routing
    - Block and transaction queries
  version: 1.0.0
  contact:
    name: OrdFS Server
servers:
  - url: /
    description: Current server
paths:
...
  /v2/tx/{txid}/proof:
    get:
      summary: Get merkle proof for transaction
      description: Returns merkle proof for transaction verification
      operationId: getMerkleProof
      tags:
        - V2 API
      parameters:
        - name: txid
          in: path
          required: true
          description: Transaction ID (64 hex characters)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{64}$'
      responses:
        '200':
          description: Merkle proof data
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '501':
          description: Not implemented yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v2/tx/{txid}/beef:
    get:
      summary: Get BEEF data for transaction
      description: Returns Background Evaluation Extended Format (BEEF) data with all transaction dependencies
      operationId: getBeef
      tags:
        - V2 API
      parameters:
        - name: txid
          in: path
          required: true
          description: Transaction ID (64 hex characters)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{64}$'
      responses:
        '200':
          description: BEEF formatted transaction data
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '501':
          description: Not implemented yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v2/tx/{txid}/{outputIndex}:
    get:
      summary: Get specific transaction output
      description: Returns raw binary data for a specific transaction output
      operationId: getOutput
      tags:
        - V2 API
      parameters:
        - name: txid
          in: path
          required: true
          description: Transaction ID (64 hex characters)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{64}$'
        - name: outputIndex
          in: path
          required: true
          description: Output index (vout)
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Raw output data
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '501':
          description: Not implemented yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v2/block/{hashOrHeight}:
    get:
      summary: Get block header in binary format
      description: Returns raw binary block header (80 bytes). Accepts either block hash (64 hex characters) or block height (integer).
      operationId: getBlockHeader
      tags:
        - V2 API
      parameters:
        - name: hashOrHeight
          in: path
          required: true
          description: Block hash (64 hex characters) or block height (integer)
          schema:
            oneOf:
              - type: string
                pattern: '^[0-9a-fA-F]{64}$'
              - type: integer
                minimum: 0
          examples:
            byHash:
              value: "000000000000000000d5c78530c2e25e8b8b8e2e2a8f8e8e2e2a8f8e8e2e2a8f"
              summary: Query by block hash
            byHeight:
              value: 800000
              summary: Query by block height
      responses:
        '200':
          description: Raw binary block header (80 bytes)
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid hash or height format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Block header not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...